+++ crates/ferroid/src/id.rs
@@ -1,6 +1,6 @@
 use crate::{Error, Result};
-use core::fmt;
-use std::{
+use core::{
+    fmt,
     hash::Hash,
     ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Sub, SubAssign},
 };
@@ -8,7 +8,7 @@ use std::{
 /// Trait for converting numeric-like values into a `u64`.
 ///
 /// This is typically used to normalize custom duration types into milliseconds
-/// for compatibility with APIs like [`std::time::Duration::from_millis`], which
+/// for compatibility with APIs like [`core::time::Duration::from_millis`], which
 /// are commonly required in async sleep contexts such as
 /// [`tokio::time::sleep`].
 pub trait ToU64 {
